+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/effect/SoundEffect.java
@@ -3,9 +3,11 @@ package com.elmakers.mine.bukkit.api.effect;
 import org.bukkit.Location;
 import org.bukkit.Sound;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
 public interface SoundEffect {
+    boolean isCustom();
     String getCustomSound();
     Sound getSound();
     float getVolume();
@@ -13,4 +15,5 @@ public interface SoundEffect {
     int getRange();
     void play(Plugin controller, Entity entity);
     void play(Plugin controller, Location location);
+    void stop(Player player);
 }