+++ Magic/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Support infinite, infinity and forever as Repeat parameters
  - Put in some fixes for wall-clipping Rocket spell exploits
  - Fix error when using a graduation NPC in potter configs without holding a wand
+ - Add StopSound action
 
 # 7.9.8